migrate: hookを@/hooksに移行
