+++ packages/akane-next/src/hooks/LocationHook.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { LocationContext } from '@/context/LocationContext';
+
+export const useLocation = () => {
+    const context = useContext(LocationContext);
+    if (!context) {
+        throw new Error('useLocation must be used within a LocationProvider');
+    }
+    return context;
+};
